+++ components/proposal/ProposalVotingList.tsx
@@ -166,7 +166,7 @@ const ProposalVotingList = () => {
     })(Tabs);
 
     return (
-        <Grid container spacing={1} justify="center" className={classes.container}>
+        <Grid container  justify="center" className={classes.container}>
             <Paper className={classes.paper}>
                 <Typography
                     color="textPrimary"