+++ pages/account/[account].tsx
@@ -10,10 +10,7 @@ const useStyles = makeStyles((theme: Theme) =>
     root: {
       display: "flex",
       justifyContent: "center",
-      paddingTop: "2%",
     },
-
-
   })
 );