+++ pages/proposal/[proposal].tsx
@@ -15,12 +15,6 @@ const useStyles = makeStyles((theme: Theme) =>
       justifyContent: "center",
     },
 
-    bottomPadding: {
-      overflow: "auto",
-      // paddingTop: '1.5%',
-      // paddingBottom: '1.5%',
-      padding: "1rem",
-    },
   })
 );
 
@@ -31,17 +25,17 @@ export default function Proposal() {
   // const pid = parseInt(router.query.block);
   return (
     <Layout>
-      <Grid container className={classes.root}>
-        <Grid item xs={12} sm={8} className={classes.bottomPadding}>
+      <Grid container spacing={2} className={classes.root}>
+        <Grid item xs={12} sm={8} >
           <MarketCard />
         </Grid>
-        <Grid item xs={12} sm={8} className={classes.bottomPadding}>
+        <Grid item xs={12} sm={8} >
           <ProposalDetails />
         </Grid>
-        <Grid item xs={12} sm={8} className={classes.bottomPadding}>
+        <Grid item xs={12} sm={8} >
           <ProposalVotingList />
         </Grid>
-        <Grid item xs={12} sm={8} className={classes.bottomPadding}>
+        <Grid item xs={12} sm={8} >
           <DepositList />
         </Grid>
       </Grid>