+++ pages/transaction/[transaction].tsx
@@ -10,13 +10,9 @@ const useStyles = makeStyles((theme: Theme) =>
     root: {
       display: "flex",
       justifyContent: "center",
-      paddingTop: "2%",
+      // paddingTop: "2%",
     },
 
-    bottomPadding: {
-      overflow: "auto",
-      padding: "1rem",
-    },
   })
 );
 
@@ -28,7 +24,7 @@ export default function Transaction() {
   return (
     <Layout>
       <Grid container className={classes.root}>
-        <Grid item xs={12} sm={8} className={classes.bottomPadding}>
+        <Grid item xs={12} sm={8}>
           <TransactionDetails hashValue={router.query.transaction} />
         </Grid>
       </Grid>